+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 
 - Add `KEDA_HTTP_DEFAULT_TIMEOUT` support in operator ([#1548](https://github.com/kedacore/keda/issues/1548))
 - Removed `MIN field` for scaledjob.([#1553](https://github.com/kedacore/keda/pull/1553))
+- Fix a memory leak in kafka client and close push scalers ([#1565](https://github.com/kedacore/keda/issues/1565))
 
 ### Breaking Changes